Simplify layout of the room invite page
Put the most important information needed to join the room leftmost.

Replace URL link with a first-time user friendly "Use {app}" button.
Don't show the invite URL which can be complex-looking.

Show each of the links only at a single place, to avoid confusion.  Move
the app homepage link to appear together with the app description, so
that where the link goes is clearer.

Don't show the URL at the bottom of the page: the user can find it in
the URL bar if necessary.

For invites, deal with the fact that the description may be missing.