fix: Update footer URL path to existing branch
In the footer there's a link to add clients to the tree of this
repository, which pointed to a branch that did not exist in the
repository anymore. As such GitHub rendered a 404. This change points to
the footer URL to the `main` branch, which should exist as it's the
default branch for this project.