Full meta
The Global Network Council (“GNC”) Representative is an individual who serves as representative of the CC Chapter on the Global Network Council. This individual must be an Individual Member of the CC Global Network and a participant in the CC Chapter. This individual is elected via processes established by the Chapter and consistent with the Chapter Guidelines. They are elected for a period, likely 1 or 2 years, to be determined at a later date.